refactor: UserRegisterRequest DTO 패키지 이동 및 Validation 제거
- DTO 패키지 변경: kr.sparta.livechat.dto -> kr.sparta.livechat.dto.user
- 비밀번호 예외 처리 로직 서비스에서 처리

Refs:#1 (GIT/Issue Tracker ID)
